# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen API",
        "description": "API for the Lumen project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/mi-horario": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Obtener mi horario (alumno)",
                "operationId": "a57569eb570d206865f99a1be98fa0f6",
                "responses": {
                    "200": {
                        "description": "Horario obtenido con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VwAlumnoHorario"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener el horario"
                    }
                }
            }
        },
        "/horarios": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Obtener todos los horarios (admin)",
                "operationId": "baf9acb677d7f71f99eedc34012efee7",
                "responses": {
                    "200": {
                        "description": "Horarios obtenidos con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VwAdminHorario"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los horarios"
                    }
                }
            },
            "post": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Crear un nuevo horario",
                "operationId": "014a5cd0cdade8bc5cd96dff76e7a49d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_ciclo",
                                    "id_cat_dia",
                                    "hora_inicio",
                                    "hora_fin"
                                ],
                                "properties": {
                                    "id_ciclo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cat_dia": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00:00"
                                    },
                                    "hora_fin": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horario creado con éxito"
                    },
                    "500": {
                        "description": "Error al crear el horario"
                    }
                }
            }
        },
        "/horarios/{id}": {
            "put": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Actualizar un horario existente",
                "operationId": "d51c9cb055713a8d2f2f432b1e432dc0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_ciclo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cat_dia": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00:00"
                                    },
                                    "hora_fin": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Horario actualizado con éxito"
                    },
                    "500": {
                        "description": "Error al actualizar el horario"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Eliminar un horario",
                "operationId": "62829f2be79ef3e6a884d6034f3985f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horario eliminado con éxito"
                    },
                    "500": {
                        "description": "Error al eliminar el horario"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VwAdminHorario": {
                "title": "VwAdminHorario",
                "description": "Vista de Horarios para Administrador",
                "properties": {
                    "id_horario": {
                        "title": "ID del Horario",
                        "description": "Identificador único del horario",
                        "type": "integer",
                        "example": 1
                    },
                    "id_ciclo": {
                        "title": "ID del Ciclo",
                        "description": "Identificador único del ciclo",
                        "type": "integer",
                        "example": 1
                    },
                    "ciclo": {
                        "title": "Ciclo",
                        "description": "Nombre del ciclo",
                        "type": "string",
                        "example": "2024-1"
                    },
                    "id_cat_dia": {
                        "title": "ID del Día",
                        "description": "Identificador único del día de la semana",
                        "type": "integer",
                        "example": 1
                    },
                    "dia": {
                        "title": "Día",
                        "description": "Día de la semana",
                        "type": "string",
                        "example": "Lunes"
                    },
                    "hora_inicio": {
                        "title": "Hora de Inicio",
                        "description": "Hora de inicio de la clase",
                        "type": "string",
                        "format": "time",
                        "example": "08:00:00"
                    },
                    "hora_fin": {
                        "title": "Hora de Fin",
                        "description": "Hora de fin de la clase",
                        "type": "string",
                        "format": "time",
                        "example": "10:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "VwAdminHorario"
                }
            },
...
            "VwAlumnoHorario": {
                "title": "VwAlumnoHorario",
                "description": "Vista de Horario de Alumno",
                "properties": {
                    "id_alumno": {
                        "title": "ID del Alumno",
                        "description": "Identificador único del alumno",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_alumno": {
                        "title": "Nombre del Alumno",
                        "description": "Nombre completo del alumno",
                        "type": "string",
                        "example": "Juan Perez"
                    },
                    "id_grupo": {
                        "title": "ID del Grupo",
                        "description": "Identificador único del grupo",
                        "type": "integer",
                        "example": 1
                    },
                    "grupo": {
                        "title": "Grupo",
                        "description": "Nombre del grupo",
                        "type": "string",
                        "example": "G-01"
                    },
                    "id_materia": {
                        "title": "ID de la Materia",
                        "description": "Identificador único de la materia",
                        "type": "integer",
                        "example": 1
                    },
                    "materia": {
                        "title": "Materia",
                        "description": "Nombre de la materia",
                        "type": "string",
                        "example": "Matemáticas"
                    },
                    "id_docente": {
                        "title": "ID del Docente",
                        "description": "Identificador único del docente",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_docente": {
                        "title": "Nombre del Docente",
                        "description": "Nombre completo del docente",
                        "type": "string",
                        "example": "Dr. Juan Perez"
                    },
                    "hora_inicio": {
                        "title": "Hora de Inicio",
                        "description": "Hora de inicio de la clase",
                        "type": "string",
                        "format": "time",
                        "example": "08:00:00"
                    },
                    "hora_fin": {
                        "title": "Hora de Fin",
                        "description": "Hora de fin de la clase",
                        "type": "string",
                        "format": "time",
                        "example": "10:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "VwAlumnoHorario"
                }
            },
...
        }
    },
...
}